Adds updateClassByBestReferenceScore, for multref alignment to set the class assignment to the best reference score, which may be useful when a subset of tomos is selected for the reference.
Tweaks to how alignRaw splits up the alignment into chunks, to make it more robust to large numbers of tomos. Still not quite right though.

Other adjustments to make flgMultiRefAlignment work, in particular, force flgAllClasses to true so that the case of a subset as refs, all subtomos are still aligned.

Optimization: in mexCTF the value was written to global memory and then if sqrCTF was true, it would square that value, requiring an extra trip through mem. Maybe the compiler would have sorted that, i did not look at the saas.

Set pca binns for diffMap to take the x-axis of the radial grid as input, achieving finer sampling for the rescaling approach.

Some small changes to handle fewer refs than class assignments when running tomoCPR and generating an average tomo with color by class.